modified:   kmean/dataset/testdata/cluster/initial 	modified:   kmean/src/mapreduce/CanopyMapper.java 	modified:   kmean/src/mapreduce/KmeansDriver.java 	modified:   kmean/src/mapreduce/KmeansMapper.java 	new file:   kmean/src/tests/AllTests.java

+++ kmean/src/mapreduce/CanopyMapper.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 
 import org.apache.hadoop.io.IntWritable;
-import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
@@ -13,7 +12,7 @@ import canopy.Canopy;
 import canopy.CanopyClusterer;
 
 public class CanopyMapper extends
-		Mapper<LongWritable, Text, IntWritable, VectorDoubleWritable> {
+		Mapper<IntWritable, Text, IntWritable, VectorDoubleWritable> {
 	private VectorDoubleWritable point = null;
 	protected CanopyClusterer canopyClusterer = new CanopyClusterer();
 	private ArrayList<Canopy> canopies = new ArrayList<Canopy>();
@@ -21,7 +20,7 @@ public class CanopyMapper extends
 
 
 	@Override
-	protected void map(LongWritable key, Text values, Context context)
+	protected void map(IntWritable key, Text values, Context context)
 			throws IOException, InterruptedException {
 		point = new VectorDoubleWritable(values);
 		canopyClusterer.addPointToCanopies(point, canopies);
